fastapi-prefect/requirements.txt
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
apprise==1.7.2
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coolname==2.2.0
croniter==2.0.1
cryptography==42.0.2
dateparser==1.2.0
dnspython==2.5.0
docker==6.1.3
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fsspec==2024.2.0
google-auth==2.27.0
graphviz==0.20.1
greenlet==3.0.3
griffe==0.40.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.9.13
packaging==23.2
pathspec==0.12.1
pendulum==2.1.2
prefect==2.14.21
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
text-unidecode==1.3
toml==0.10.2
typer==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.0
uvicorn==0.27.1
websocket-client==1.7.0
websockets==12.0
